drop old tests
